Moved netcoreapp specific tests to a new file so that tests targeting netfx builds
